Add stop and getQueueInfo methods
stop(true) will stop any serviceQueue() loops that are running
when there are no tasks left on the queue.
stop(false) will stop any serviceQueue() loops as soon as they've
finished any task that they might be in the middle of servicing.

getQueueInfo returns the number of tasks on the queue and,
if there are tasks, scheduled times of first and last tasks.